requirements.txt
Django==2.0.3
stripe-2.41.0